fixed: Ticket #8954 - XBMC crashes when playing a particular FLAC file
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@28442 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/paplayer/FLACcodec.cpp
@@ -99,7 +99,7 @@ bool FLACCodec::Init(const CStdString &strFile, unsigned int filecache)
   if (tag.Read(strFile))
     m_replayGain=tag.GetReplayGain();
 
-  m_Bitrate = (int)((m_file.GetLength() * 8) / (m_TotalTime/1000));
+  m_Bitrate = (int)(((float)m_file.GetLength() * 8) / ((float)m_TotalTime / 1000));
 
   if (m_pBuffer)
   {
